+++ features/unmute.go
@@ -1,0 +1,12 @@
+package features
+
+// dg "github.com/bwmarrin/discordgo"
+
+// 												***		U N M U T E 	***
+
+//
+/*
+func Unmute(s *dg.Session, m *dg.MessageCreate) {
+
+}
+*/
