+++ go.mod
@@ -2,4 +2,7 @@ module github.com/notAxion/HackerHeads
 
 go 1.16
 
-require github.com/bwmarrin/discordgo v0.23.2
+require (
+	github.com/bwmarrin/discordgo v0.23.2
+	github.com/qiniu/qmgo v0.9.3
+)
